NXP-23132: move nuxeo-core-kv to nuxeo-runtime-kv
