+++ nuxeo-core/nuxeo-core-api/pom.xml
@@ -50,8 +50,8 @@
       <artifactId>nuxeo-runtime-jtajca</artifactId>
     </dependency>
     <dependency>
-      <groupId>org.nuxeo.ecm.core</groupId>
-      <artifactId>nuxeo-core-kv</artifactId>
+      <groupId>org.nuxeo.runtime</groupId>
+      <artifactId>nuxeo-runtime-kv</artifactId>
     </dependency>
     <dependency>
       <groupId>javax.servlet</groupId>
