--- nuxeo-core/nuxeo-core-kv/src/main/resources/META-INF/MANIFEST.MF
@@ -1,4 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-SymbolicName: org.nuxeo.ecm.core.kv
-Nuxeo-Component: OSGI-INF/keyvalue-service.xml
