--- nuxeo-core/nuxeo-core-kv/src/test/resources/META-INF/MANIFEST.MF
@@ -1,3 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-SymbolicName: org.nuxeo.ecm.core.kv.tests
