+++ nuxeo-core/nuxeo-core-lock/pom.xml
@@ -13,12 +13,12 @@
 
   <dependencies>
     <dependency>
-      <groupId>org.nuxeo.ecm.core</groupId>
-      <artifactId>nuxeo-core-api</artifactId>
+      <groupId>org.nuxeo.runtime</groupId>
+      <artifactId>nuxeo-runtime-kv</artifactId>
     </dependency>
     <dependency>
       <groupId>org.nuxeo.ecm.core</groupId>
-      <artifactId>nuxeo-core-kv</artifactId>
+      <artifactId>nuxeo-core-api</artifactId>
     </dependency>
   </dependencies>
