+++ nuxeo-core/nuxeo-core-mongodb/pom.xml
@@ -33,8 +33,8 @@
       <artifactId>nuxeo-core-storage</artifactId>
     </dependency>
     <dependency>
-      <groupId>org.nuxeo.ecm.core</groupId>
-      <artifactId>nuxeo-core-kv</artifactId>
+      <groupId>org.nuxeo.runtime</groupId>
+      <artifactId>nuxeo-runtime-kv</artifactId>
     </dependency>
 
     <dependency>
@@ -64,8 +64,8 @@
       <scope>test</scope>
     </dependency>
     <dependency>
-      <groupId>org.nuxeo.ecm.core</groupId>
-      <artifactId>nuxeo-core-kv</artifactId>
+      <groupId>org.nuxeo.runtime</groupId>
+      <artifactId>nuxeo-runtime-kv</artifactId>
       <type>test-jar</type>
       <scope>test</scope>
     </dependency>
