+++ nuxeo-core/pom.xml
@@ -24,7 +24,6 @@
     <module>nuxeo-core-query</module>
     <module>nuxeo-core-schema</module>
     <module>nuxeo-core-io</module>
-    <module>nuxeo-core-kv</module>
     <module>nuxeo-core-storage</module>
     <module>nuxeo-core-storage-sql</module>
     <module>nuxeo-core-convert-api</module>
