+++ nuxeo-distribution/nuxeo-nxr-server/src/main/resources/templates/common-base/nxserver/config/redis-config.xml.nxftl
@@ -42,7 +42,7 @@
 </#if>
 
 <#if "${nuxeo.keyvalue.provider}" == "redis">
-  <extension target="org.nuxeo.ecm.core.kv.KeyValueService" point="configuration">
+  <extension target="org.nuxeo.runtime.kv.KeyValueService" point="configuration">
     <store name="default" class="org.nuxeo.ecm.core.redis.contribs.RedisKeyValueStore">
       <property name="namespace">kv</property>
     </store>
