+++ nuxeo-runtime/nuxeo-runtime-kv/src/main/resources/META-INF/MANIFEST.MF
@@ -1,0 +1,4 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-SymbolicName: org.nuxeo.runtime.kv
+Nuxeo-Component: OSGI-INF/keyvalue-service.xml
