+++ nuxeo-runtime/nuxeo-runtime-kv/src/main/resources/OSGI-INF/keyvalue-service.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0"?>
+<component name="org.nuxeo.runtime.kv.KeyValueService" version="1.0">
+
+  <documentation>
+    The Key/Value service allows registration and access to KeyValueStores
+    to store simple values associated to keys.
+  </documentation>
+
+  <service>
+    <provide interface="org.nuxeo.runtime.kv.KeyValueService" />
+  </service>
+
+  <implementation class="org.nuxeo.runtime.kv.KeyValueServiceImpl" />
+
+  <extension-point name="configuration">
+    <documentation>
+      Defines the implementations of the Key/Value stores:
+      <code>
+        <store name="default" class="org.nuxeo.runtime.kv.MemKeyValueStore"/>
+      </code>
+      The class must implement org.nuxeo.runtime.kv.KeyValueStoreProvider.
+    </documentation>
+
+    <object class="org.nuxeo.runtime.kv.KeyValueStoreDescriptor" />
+  </extension-point>
+
+</component>
