+++ nuxeo-runtime/nuxeo-runtime-kv/src/test/resources/META-INF/MANIFEST.MF
@@ -1,0 +1,3 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-SymbolicName: org.nuxeo.runtime.kv.tests
