+++ nuxeo-runtime/pom.xml
@@ -27,6 +27,7 @@
     <module>nuxeo-runtime-metrics</module>
     <module>nuxeo-runtime-jtajca</module>
     <module>nuxeo-runtime-datasource</module>
+    <module>nuxeo-runtime-kv</module>
     <module>nuxeo-runtime-test</module>
     <module>nuxeo-launcher-commons</module>
     <module>nuxeo-connect-standalone</module>
